backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas-ta
google-genai
python-dotenv
httpx
feedparser
pandas
numpy
slowapi
